Expansion and bug-fixes - CAREFUL! The pipeline includes defaults which are optimised for codon.
extract_sequences.py: Remove the newline at the end of the extracted sequences.

pipeline.py: Expand it with all the latest modules/scripts. All the settings are the default.

process_blast_output.py: Write the function that writes the parsed hits to a JSON file.

process_results.py: Important bug fix -> The sequence marked for extraction in the case of
[reverse 5 > 50bp] was wrong. Fix it and add relevant comments as well.